+++ Tests/LinuxMain.swift
@@ -45,6 +45,12 @@ extension BlockBasedKVORuleTests {
     ]
 }
 
+extension CaptureVariableRuleTests {
+    static var allTests: [(String, (CaptureVariableRuleTests) -> () throws -> Void)] = [
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+    ]
+}
+
 extension ClassDelegateProtocolRuleTests {
     static var allTests: [(String, (ClassDelegateProtocolRuleTests) -> () throws -> Void)] = [
         ("testWithDefaultConfiguration", testWithDefaultConfiguration)
@@ -1798,6 +1804,7 @@ XCTMain([
     testCase(AttributesRuleTests.allTests),
     testCase(BalancedXCTestLifecycleRuleTests.allTests),
     testCase(BlockBasedKVORuleTests.allTests),
+    testCase(CaptureVariableRuleTests.allTests),
     testCase(ClassDelegateProtocolRuleTests.allTests),
     testCase(ClosingBraceRuleTests.allTests),
     testCase(ClosureBodyLengthRuleTests.allTests),